core: Lower `SigningBackend` and `Backend` with its related utilities
This moves the first interface into the new core library. This also moves the `Signer` struct by 
dropping the `Config` dependency to preserve the `jj-lib` API it is imported as `CoreSigner`.

Part of #6284
