+++ core/Cargo.toml
@@ -17,8 +17,11 @@ name = "diff_bench"
 harness = false
 
 [dependencies]
+async-trait = { workspace = true }
 blake2 = { workspace = true }
 bstr = { workspace = true }
+chrono = { workspace = true }
+clru = { workspace = true }
 digest = { workspace = true }
 either = { workspace = true }
 etcetera = { workspace = true }
